Fix errors from invalid merge with master.
