--- src/main/scala/EShop/lab4/TypedPersistentCheckout.scala
@@ -1,65 +1,0 @@
-package EShop.lab4
-
-import EShop.lab2.TypedCartActor
-import EShop.lab3.{TypedOrderManager, TypedPayment}
-import akka.actor.Cancellable
-import akka.actor.typed.scaladsl.{ActorContext, Behaviors}
-import akka.actor.typed.{ActorRef, Behavior}
-import akka.persistence.typed.PersistenceId
-import akka.persistence.typed.scaladsl.{Effect, EventSourcedBehavior}
-
-import scala.concurrent.duration._
-
-class TypedPersistentCheckout {
-
-  import EShop.lab2.TypedCheckout._
-
-  val timerDuration: FiniteDuration = 1.seconds
-
-  def schedule(context: ActorContext[Command]): Cancellable = ???
-
-  def apply(cartActor: ActorRef[TypedCartActor.Command], persistenceId: PersistenceId): Behavior[Command] =
-    Behaviors.setup { context =>
-      EventSourcedBehavior(
-        persistenceId,
-        WaitingForStart,
-        commandHandler(context, cartActor),
-        eventHandler(context)
-      )
-    }
-
-  def commandHandler(
-    context: ActorContext[Command],
-    cartActor: ActorRef[TypedCartActor.Command]
-  ): (State, Command) => Effect[Event, State] = (state, command) => {
-    state match {
-      case WaitingForStart =>
-        ???
-
-      case SelectingDelivery(_) =>
-        ???
-
-      case SelectingPaymentMethod(_) =>
-        ???
-
-      case ProcessingPayment(_) =>
-        ???
-
-      case Cancelled =>
-        ???
-
-      case Closed =>
-        ???
-    }
-  }
-
-  def eventHandler(context: ActorContext[Command]): (State, Event) => State = (state, event) => {
-    event match {
-      case CheckoutStarted           => ???
-      case DeliveryMethodSelected(_) => ???
-      case PaymentStarted(_)         => ???
-      case CheckOutClosed            => ???
-      case CheckoutCancelled         => ???
-    }
-  }
-}
